Stain normalization: white point + bit-depth handling (step 2/5)
- Default white point is now dtype-aware: dtype_max() gives the full-white
  value (255 / 65535 / 1.0) and default_white_point() raises with guidance when
  the data clearly doesn't match its dtype's range (8-bit-in-uint16, 0-255 float).
- Bit-depth-agnostic reconstruction: sda_to_rgb / lab_ruderman_to_rgb take an
  out_dtype and clip to that dtype's valid range (dtype_max) rather than a
  hardcoded 255 - threaded from the source image dtype through apply. Per review,
  the dtype (not a derived max_value float) is the threaded parameter.
- estimate_white_point is now sdata-level and samples the per-channel MEDIAN over
  non-tissue (background) pixels via the tissue mask (HistomicsTK semantics),
  replacing the whole-image percentile that under-estimated on dim scans.
- Renamed _background.py docstring/semantics to white-point; test fixtures are
  uint8 (real H&E) rather than float-0-255 (which the [0,1]-float convention
  would flag).

125 stain tests pass.

+++ tests/experimental/test_stain_normalize.py
@@ -37,7 +37,7 @@ def _make_sdata(
 @pytest.fixture
 def rgb_values() -> np.ndarray:
     rng = np.random.default_rng(3)
-    return rng.uniform(40.0, 200.0, size=(3, 64, 64)).astype(np.float32)
+    return rng.uniform(40.0, 200.0, size=(3, 64, 64)).astype(np.uint8)
 
 
 class TestFitStainReference:
